Python2ForStreamingAnalytics/requirements.txt
cycler==0.10.0
matplotlib==1.5.3
numpy==1.11.3
pandas==0.19.2
Pillow==4.0.0
pyparsing==2.1.4
python-dateutil==2.6.0
pytz==2016.10
scipy==0.18.1
six==1.10.0
